Backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
pandas>=2.2.0
anthropic>=0.25.7
requests>=2.31.0
requests-oauthlib>=1.3.1
python-jose[cryptography]>=3.3.0
pydantic>=2.6.4
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4